packages: Add patch to execute fzf history directly
The default behavior of fzf's CTRL-R feature is to paste the selected
command from the history into the command line. I would like to execute
directly instead when Enter is pressed and only paste it to the command
line when the left arrow key is pressed instead - similar to bash's and
zsh's default CTRL-R behavior.